[#141] refactor : 하이드레이션 오류 해결 및 코드 리팩토링
