+++ app/onboarding/page.tsx
@@ -1,10 +1,10 @@
 "use client";
 
 import { useState } from "react";
-import * as styles from "@/app/onboarding/page.css";
+import * as styles from "./style.css";
 import Onboarding from "@/app/onboarding/_components/OnbordingUI";
-import { useRouter } from "next/navigation";
 import OnboardingData from "@/app/onboarding/_components/OnboardingData";
+import { useRouter } from "next/navigation";
 
 const Page = () => {
   const router = useRouter();
